+++ apps/api/src/pkg/db.ts
@@ -2,12 +2,15 @@ import { type PlanetScaleDatabase, drizzle } from "drizzle-orm/planetscale-serve
 
 import { Client } from "@planetscale/database";
 import { schema } from "@unkey/db";
+import type { Logger } from "./logging";
 export type Database = PlanetScaleDatabase<typeof schema>;
 
 type ConnectionOptions = {
   host: string;
   username: string;
   password: string;
+  retry: number | false;
+  logger?: Logger;
 };
 
 export function createConnection(opts: ConnectionOptions): Database {
@@ -20,11 +23,33 @@ export function createConnection(opts: ConnectionOptions): Database {
       fetch: (url: string, init: any) => {
         (init as any).cache = undefined; // Remove cache header
         const u = new URL(url);
-        // set protocol to http if localhost for CI testing
-        if (u.host.includes("localhost")) {
+        /**
+         * Running workerd in docker caused an issue where it was trying to use https but
+         * encountered an ssl version error
+         *
+         * This enforces the use of http
+         */
+        if (u.hostname === "planetscale" || u.host.includes("localhost")) {
           u.protocol = "http";
         }
-        return fetch(u, init);
+
+        if (!opts.retry) {
+          return fetch(u, init);
+        }
+
+        let err: Error | undefined = undefined;
+        for (let i = 0; i <= opts.retry; i++) {
+          try {
+            return fetch(u, init);
+          } catch (e) {
+            opts.logger?.warn("fetching from planetscale failed", {
+              url: u.toString(),
+              attempt: i + 1,
+            });
+            err = e as Error;
+          }
+        }
+        throw err;
       },
     }),
     {