+++ apps/semantic-cache/src/analytics.ts
@@ -2,18 +2,16 @@ import { Tinybird } from "@chronark/zod-bird";
 import { z } from "zod";
 
 export const eventSchema = z.object({
-  time: z.number(),
+  timestamp: z.string(),
   model: z.string(),
   stream: z.boolean(),
   query: z.string(),
   vector: z.array(z.number()),
   response: z.string(),
   cache: z.boolean(),
-  latency: z.number(),
+  timing: z.number(),
   tokens: z.number(),
   requestId: z.string(),
-  workspaceId: z.string(),
-  gatewayId: z.string(),
 });
 
 export class Analytics {
@@ -27,7 +25,7 @@ export class Analytics {
 
   public get ingestLogs() {
     return this.client.buildIngestEndpoint({
-      datasource: "semantic_cache__v6",
+      datasource: "semantic_cache__v3",
       event: eventSchema,
     });
   }